Merge release/1.6 -> main

+++ client/ios/VPNExtension/InterfaceConfiguration.swift
@@ -5,7 +5,7 @@ final class InterfaceConfiguration: Codable {
     var privateKey: String
     var addresses: [IpAddrMask] = []
     var listenPort: UInt16?
-    var mtu: UInt16?
+    var mtu: UInt32?
     var dns: [String] = []
     var dnsSearch: [String] = []
